♻️ Refactor: 취소 페이지 에러 처리, 주석 추가, 필터링 부분 파일 분리
