+++ src/pages/admin/CancelMenu.jsx
@@ -11,13 +11,15 @@ import BottomSheet from '@/components/Admin/BottomSheet';
 import MenuFilterBox from '@/components/Admin/MenuFilterBox';
 import OrderReturnModal from '@/components/Admin/OrderReturnModal';
 import PastDateModal from '@/components/Admin/PastDateModal';
-import { FILTERS, isPastDate } from '@/constants/menuFilterData';
+import Toast from '@/components/common/Toast';
+import { FILTERS, filterOrders, isPastDate } from '@/constants/menuFilterData';
 
 export default function CancelMenu() {
   const [modal, setModal] = useState(null);
   const [selectedOrderId, setSelectedOrderId] = useState(null);
   const [dateFilter, setDateFilter] = useState('all');
   const [searchQuery, setSearchQuery] = useState('');
+  const [toast, setToast] = useState({ visible: false, message: '' });
   const navigate = useNavigate();
   const queryClient = useQueryClient();
   const { notifyOrderStatus, clearCount } = useOutletContext() ?? {};
@@ -35,9 +37,11 @@ export default function CancelMenu() {
   });
 
   useEffect(() => {
+    //페이지 진입시 구독
     const eventSource = subscribeOrder('CANCELED');
 
     const handleCanceledOrder = (event) => {
+      //새로운 데이터 들어올 시 추가
       const newOrder = JSON.parse(event.data);
       queryClient.setQueryData(queryKey, (prev) => ({
         ...(prev ?? {}),
@@ -46,6 +50,7 @@ export default function CancelMenu() {
     };
 
     const handleNotification = (event) => {
+      //다른 페이지 데이터 추가시 표시
       const { orderStatus } = JSON.parse(event.data);
       notifyOrderStatus?.(orderStatus);
     };
@@ -66,16 +71,7 @@ export default function CancelMenu() {
     };
   }, [queryClient, notifyOrderStatus]);
 
-  const q = searchQuery.trim().toLowerCase();
-  const filtered = orderData.filter((d) => {
-    if (dateFilter !== 'all' && d.orderDate !== dateFilter) return false;
-    if (!q) return true;
-    return (
-      d.customerName.toLowerCase().includes(q) ||
-      String(d.tableNumber).includes(q) ||
-      d.customerPhoneNumber.includes(q)
-    );
-  });
+  const filtered = filterOrders(orderData, dateFilter, searchQuery); //필터링된 데이터
 
   const closeModal = () => {
     setModal(null);
@@ -83,15 +79,17 @@ export default function CancelMenu() {
   };
 
   const handleUndoClick = (order) => {
+    //되돌리기 버튼 클릭 시
     setSelectedOrderId(order.orderId);
     if (isPastDate(order.orderDate)) {
-      setModal('pastDate');
+      setModal('pastDate'); //지난 날짜면 불가능
     } else {
       setModal('undoConfirm');
     }
   };
 
   const handleUndoConfirm = async (orderId) => {
+    //되돌리기 요청
     if (!orderId) return;
     try {
       await patchChangeOrderStatus(orderId, 'WAITING');
@@ -102,12 +100,13 @@ export default function CancelMenu() {
       setModal('undoDone');
     } catch (error) {
       console.log('대기로 되돌리기 실패: ' + error);
+      setToast({ visible: true, message: '잠시후 다시 시도해주세요', icon: 'warning' });
     }
   };
 
   return (
     <div className="flex flex-col w-full h-full bg-[#F8F8F8] items-center">
-      <MenuFilterBox
+      <MenuFilterBox /* 주문 필터링 박스 */
         title="취소된 주문"
         count={filtered.length}
         filters={FILTERS}
@@ -120,7 +119,7 @@ export default function CancelMenu() {
       {filtered.length > 0 ? (
         <div className="flex flex-col w-full gap-2 overflow-auto no-scrollbar px-5 py-5">
           {filtered.map((data) => (
-            <CompletedOrderCard
+            <CompletedOrderCard /* 취소 주문 카드 */
               key={data.orderId}
               tableNumber={data.tableNumber}
               peopleCount={data.numOfPeople}
@@ -147,7 +146,7 @@ export default function CancelMenu() {
         </div>
       )}
 
-      <BottomSheet
+      <BottomSheet /* 대기로 되돌리기 확인 모달 박스 */
         open={modal === 'undoConfirm'}
         onOpenChange={(o) => !o && closeModal()}
         showButton
@@ -163,7 +162,7 @@ export default function CancelMenu() {
         </div>
       </BottomSheet>
 
-      <OrderReturnModal
+      <OrderReturnModal /* 주문 되돌리기 완료 확인 박스 */
         open={modal === 'undoDone'}
         onOpenChange={(o) => !o && closeModal()}
         onMove={() => {
@@ -173,10 +172,17 @@ export default function CancelMenu() {
         onConfirm={closeModal}
       />
 
-      <PastDateModal
+      <PastDateModal /* 지난 날짜 되돌리기 불가 안내 모달 */
         open={modal === 'pastDate'}
         onOpenChange={(o) => !o && closeModal()}
         onConfirm={closeModal}
+      />
+
+      <Toast /* 요청 실패시 재시도 안내 토스트 */
+        visible={toast.visible}
+        message={toast.message}
+        icon={toast.icon ?? 'check'}
+        onClose={() => setToast({ visible: false, message: '' })}
       />
     </div>
   );
